documented and refactored runSync
