+++ apps/cloud/app/(auth)/layout.tsx
@@ -1,6 +1,8 @@
 import { redirect } from "next/navigation";
 import { getSupabaseServerClientReadOnly } from "@/api/supabase/server-readonly";
 
+// Auth pages are only for signed-out users, so this layout redirects existing
+// sessions away from /login before rendering the login form.
 export default async function AuthLayout({
   children,
 }: {
